+++ Models/Nodes/GuideNode.cs
@@ -7,7 +7,7 @@
 
         public override NodeType Type => NodeType.GuideNode;
 
-        public GuideNode(double x, double y, string name = null) :
+        public GuideNode(int x, int y, string name = null) :
             base(x, y)
         {
             Name = name;
